Fix code according to reviews from zhhsplendid
